requirements.txt
pandas==2.2.3
pillow==11.0.0
pip==25.0.1
mpmath==1.3.0
numpy==2.2.4
opencv-python==4.11.0.86
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1

